+++ tools/cosmovisor/process.go
@@ -283,7 +283,7 @@ func (l Launcher) WaitForUpgradeOrExit(cmd *exec.Cmd) (bool, error) {
 		if l.cfg.ShutdownGrace > 0 {
 			// Interrupt signal
 			l.logger.Info("sent interrupt to app, waiting for exit")
-			_ = cmd.Process.Signal(os.Interrupt)
+			_ = cmd.Process.Signal(syscall.SIGTERM)
 
 			// Wait app exit
 			psChan := make(chan *os.ProcessState)