+++ x/gov/simulation/operations.go
@@ -294,7 +294,7 @@ func simulateMsgSubmitProposal(
 		votingPeriod := params.VotingPeriod
 		s := NewSharedState()
 		fops := make([]simtypes.FutureOperation, numVotes+1)
-		for i := 0; i < numVotes; i++ {
+		for i := range numVotes {
 			whenVote := ctx.BlockHeader().Time.Add(time.Duration(r.Int63n(int64(votingPeriod.Seconds()))) * time.Second)
 			fops[i] = simtypes.FutureOperation{
 				BlockTime: whenVote,