+++ types/events_test.go
@@ -90,7 +90,7 @@ func (s *eventsTestSuite) TestEventManager() {
 
 func (s *eventsTestSuite) TestEmitTypedEvent() {
 	s.Run("deterministic key-value order", func() {
-		for i := 0; i < 10; i++ {
+		for range 10 {
 			em := sdk.NewEventManager()
 			coin := sdk.NewCoin("fakedenom", math.NewInt(1999999))
 			s.Require().NoError(em.EmitTypedEvent(&coin))