+++ types/module/module_int_test.go
@@ -17,7 +17,7 @@ type TestSuite struct {
 
 func (s *TestSuite) TestAssertNoForgottenModules() {
 	m := Manager{
-		Modules: map[string]interface{}{"a": nil, "b": nil},
+		Modules: map[string]any{"a": nil, "b": nil},
 	}
 	tcs := []struct {
 		name     string
@@ -42,7 +42,7 @@ func (s *TestSuite) TestAssertNoForgottenModules() {
 
 func (s *TestSuite) TestModuleNames() {
 	m := Manager{
-		Modules: map[string]interface{}{"a": nil, "b": nil},
+		Modules: map[string]any{"a": nil, "b": nil},
 	}
 	ms := m.ModuleNames()
 	sort.Strings(ms)