+++ client/v2/CHANGELOG.md
@@ -38,7 +38,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Features
 
-* [#18461](https://github.com/cosmos/cosmos-sdk/pull/18461) Support governance proposals.
+* [#24359](https://github.com/cosmos/cosmos-sdk/pull/24359) Support governance proposals.
 
 ### Bug Fixes