+++ client/v2/CHANGELOG.md
@@ -36,6 +36,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Features
+
+* [#24359](https://github.com/cosmos/cosmos-sdk/pull/24359) Support governance proposals.
+
 ## [v2.0.0-beta.8] - 2025-01-29
 
 * [#23544](https://github.com/cosmos/cosmos-sdk/pull/23544) Support map in queries for autocli.